--- src/fs/file.ts
@@ -1,5 +1,0 @@
-// TODO
-
-export interface File_Filter {
-	(id: string): boolean;
-}
