+++ src/fs/filter.ts
@@ -1,7 +1,6 @@
 import {join} from 'path';
 
 import {paths} from '../paths.js';
-import type {File_Filter} from 'src/fs/file.js';
 import type {Path_Stats} from 'src/fs/path_data.js';
 
 // This is a subset of the `cheap-watch` types designed for browser compatibility.
@@ -10,6 +9,14 @@ export interface Path_Filter {
 }
 
 export const to_path_filter =
-	(exclude: File_Filter, root = paths.root): Path_Filter =>
-	({path}) =>
-		!exclude(join(root, path));
+	(exclude: Id_Stats_Filter, root = paths.root): Path_Filter =>
+	({path, stats}) =>
+		!exclude(join(root, path), stats);
+
+export interface Id_Stats_Filter {
+	(id: string, stats: Path_Stats): boolean;
+}
+
+export interface Id_Filter {
+	(id: string): boolean;
+}
