+++ src/fs/node.ts
@@ -4,7 +4,7 @@ import {sort_map, compare_simple_map_entries} from '@feltcoop/felt/util/map.js';
 
 import type {Filesystem, Fs_Write_File} from 'src/fs/filesystem.js';
 import type {Path_Stats} from 'src/fs/path_data.js';
-import type {Path_Filter} from 'src/fs/path_filter.js';
+import type {Path_Filter} from 'src/fs/filter.js';
 
 // This uses `Cheap_Watch` which probably isn't the fastest, but it works fine for now.
 // TODO should this API be changed to only include files and not directories?
